feat: more retries for checking commit status (#1275)
## Summary


Try to mitigate the issue seen: 


https://console.cloud.google.com/dataproc/jobs/71ce070e-2505-479b-ae09-7092b3094e7f/monitoring?region=us-central1&project=canary-443022

```bash
Cannot determine whether the commit was successful or not, the underlying data files may or may not be needed. Manual intervention via the Remove Orphan Files Action can remove these files when a connection to the Catalog can be re-established if the commit was actually unsuccessful.
Please check to see whether or not your commit was successful before retrying this commit. Retrying an already successful operation will result in duplicate records or unintentional modifications.
At this time no files will be deleted including possibly unused manifest lists.
```


## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update


<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

## Chores
- Improved transaction processing reliability by enhancing commit status
verification configuration with optimized retry attempts and wait
intervals, providing better resilience when handling distributed system
delays and transient failures.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

Co-authored-by: thomaschow <thomaschow369@gmail.com>