Update our required triggered checks to also work for fork based PRs (#1319)
## Summary
Currently our triggered checks work on 'pushes' to branches in our repo.
This seems to work ok for internal contribs but seems to not do the
trick for PRs submitted via forks (e.g.
https://github.com/zipline-ai/chronon/pull/1312). Updating the triggers
to also work on pull_requests to see if this fixes our CI status checks
on that PR.

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

* **Chores**
* Updated CI workflow triggers to run only for activity on the main
branch (pushes, pull requests targeting main, and merge group events),
narrowing the scope of automated checks for improved efficiency and
reduced noise.

<sub>✏️ Tip: You can customize this high-level summary in your review
settings.</sub>
<!-- end of auto-generated comment: release notes by coderabbit.ai -->